+++ .gitignore
@@ -1,8 +1,5 @@
 target/
 
-# UnQLite database created during cargo test
-**/*.unql
-
 # These are backup files generated by rustfmt
 **/*.rs.bk
