+++ Cargo.toml
@@ -23,7 +23,6 @@ systemd-units = { unit-name = "pyrsia", unit-scripts = "installers/debian/" }
 [dependencies]
 pyrsia_blockchain_network = { path = "src/blockchain" }
 
-
 anyhow = "1.0.57"
 async-trait = "0.1.53"
 base64 = "0.13.0"
@@ -37,7 +36,6 @@ clap = { version = "3.1.18", features = ["derive"] }
 dirs = "4.0.0"
 defaults = "0.2.0"
 derive_builder = "0.11.2"
-easy-hasher = "2.2.1"
 env_logger = "0.9.0"
 filename = "0.1.1"
 fs_extra = "1.2.0"
@@ -70,12 +68,10 @@ test-log = "0.2.8"
 thiserror = "1.0.31"
 tokio = { version = "1.18.2", features = [ "macros", "rt-multi-thread", "io-std" ] }
 toml = "0.5.9"
-unqlite = "1.5.0"
 url = "2.2.2"
 uuid = { version = "1.1.1", features = [ "v4" ] }
 warp = { version = "0.3.1", default-features = false }
 walkdir = "2.3.2"
-
 
 [dependencies.error-chain]
 version = "0.12"
