+++ pyrsia_central_node/src/main.rs
@@ -19,6 +19,7 @@ pub mod network;
 
 use args::parser::PyrsiaNodeArgs;
 use network::handlers;
+use pyrsia::artifact_service::storage::ArtifactStorage;
 use pyrsia::docker::error_util::*;
 use pyrsia::docker::v2::routes::make_docker_routes;
 use pyrsia::logging::*;
@@ -50,8 +51,16 @@ async fn main() -> Result<(), Box<dyn Error>> {
     debug!("Start p2p event loop");
     tokio::spawn(event_loop.run());
 
+    debug!("Create artifact storage");
+    let artifact_storage = ArtifactStorage::new()?;
+
     debug!("Setup HTTP server");
-    setup_http(&args, transparency_log, p2p_client.clone());
+    setup_http(
+        &args,
+        transparency_log,
+        p2p_client.clone(),
+        artifact_storage.clone(),
+    );
 
     debug!("Start p2p components");
     setup_p2p(p2p_client.clone(), args).await;
@@ -68,6 +77,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
                 } => {
                     if let Err(error) = handlers::handle_request_artifact(
                         p2p_client.clone(),
+                        artifact_storage.clone(),
                         &artifact_type,
                         &artifact_hash,
                         channel,
@@ -95,7 +105,12 @@ async fn main() -> Result<(), Box<dyn Error>> {
     }
 }
 
-fn setup_http(args: &PyrsiaNodeArgs, transparency_log: TransparencyLog, p2p_client: Client) {
+fn setup_http(
+    args: &PyrsiaNodeArgs,
+    transparency_log: TransparencyLog,
+    p2p_client: Client,
+    artifact_storage: ArtifactStorage,
+) {
     // Get host and port from the settings. Defaults to DEFAULT_HOST and DEFAULT_PORT
     debug!(
         "Pyrsia Docker Node will bind to host = {}, port = {}",
@@ -108,8 +123,12 @@ fn setup_http(args: &PyrsiaNodeArgs, transparency_log: TransparencyLog, p2p_clie
     );
 
     debug!("Setup HTTP routing");
-    let docker_routes = make_docker_routes(transparency_log, p2p_client.clone());
-    let node_api_routes = make_node_routes(p2p_client);
+    let docker_routes = make_docker_routes(
+        transparency_log,
+        p2p_client.clone(),
+        artifact_storage.clone(),
+    );
+    let node_api_routes = make_node_routes(p2p_client, artifact_storage);
     let all_routes = docker_routes.or(node_api_routes);
 
     debug!("Setup HTTP server");
@@ -138,13 +157,5 @@ async fn setup_p2p(mut p2p_client: Client, args: PyrsiaNodeArgs) {
 
     if let Some(to_dial) = args.peer {
         handlers::dial_other_peer(p2p_client.clone(), &to_dial).await;
-    }
-
-    debug!("Provide local artifacts");
-    if let Err(error) = handlers::provide_artifacts(p2p_client.clone()).await {
-        warn!(
-            "An error occurred while providing local artifacts. Error: {:?}",
-            error
-        );
     }
 }
