+++ pyrsia_node/src/main.rs
@@ -20,6 +20,7 @@ pub mod network;
 use anyhow::{bail, Result};
 use args::parser::PyrsiaNodeArgs;
 use network::handlers;
+use pyrsia::artifact_service::storage::ArtifactStorage;
 use pyrsia::docker::error_util::*;
 use pyrsia::docker::v2::routes::make_docker_routes;
 use pyrsia::logging::*;
@@ -52,6 +53,9 @@ async fn main() -> Result<(), Box<dyn Error>> {
     debug!("Create p2p components");
     let (p2p_client, mut p2p_events, event_loop) = p2p::setup_libp2p_swarm(args.max_provided_keys)?;
 
+    debug!("Create artifact storage");
+    let artifact_storage = ArtifactStorage::new()?;
+
     debug!("Create blockchain components");
     let _blockchain = setup_blockchain()?;
 
@@ -59,7 +63,12 @@ async fn main() -> Result<(), Box<dyn Error>> {
     tokio::spawn(event_loop.run());
 
     debug!("Setup HTTP server");
-    setup_http(&args, transparency_log, p2p_client.clone());
+    setup_http(
+        &args,
+        transparency_log,
+        p2p_client.clone(),
+        artifact_storage.clone(),
+    );
 
     debug!("Start p2p components");
     setup_p2p(p2p_client.clone(), args).await;
@@ -76,6 +85,7 @@ async fn main() -> Result<(), Box<dyn Error>> {
                 } => {
                     if let Err(error) = handlers::handle_request_artifact(
                         p2p_client.clone(),
+                        artifact_storage.clone(),
                         &artifact_type,
                         &artifact_hash,
                         channel,
@@ -103,7 +113,12 @@ async fn main() -> Result<(), Box<dyn Error>> {
     }
 }
 
-fn setup_http(args: &PyrsiaNodeArgs, transparency_log: TransparencyLog, p2p_client: Client) {
+fn setup_http(
+    args: &PyrsiaNodeArgs,
+    transparency_log: TransparencyLog,
+    p2p_client: Client,
+    artifact_storage: ArtifactStorage,
+) {
     // Get host and port from the settings. Defaults to DEFAULT_HOST and DEFAULT_PORT
     debug!(
         "Pyrsia Docker Node will bind to host = {}, port = {}",
@@ -116,8 +131,12 @@ fn setup_http(args: &PyrsiaNodeArgs, transparency_log: TransparencyLog, p2p_clie
     );
 
     debug!("Setup HTTP routing");
-    let docker_routes = make_docker_routes(transparency_log, p2p_client.clone());
-    let node_api_routes = make_node_routes(p2p_client);
+    let docker_routes = make_docker_routes(
+        transparency_log,
+        p2p_client.clone(),
+        artifact_storage.clone(),
+    );
+    let node_api_routes = make_node_routes(p2p_client, artifact_storage);
     let all_routes = docker_routes.or(node_api_routes);
 
     debug!("Setup HTTP server");
@@ -146,13 +165,6 @@ async fn setup_p2p(mut p2p_client: Client, args: PyrsiaNodeArgs) {
 
     if let Some(to_dial) = args.peer {
         handlers::dial_other_peer(p2p_client.clone(), &to_dial).await;
-    }
-    debug!("Provide local artifacts");
-    if let Err(error) = handlers::provide_artifacts(p2p_client.clone()).await {
-        warn!(
-            "An error occurred while providing local artifacts. Error: {:?}",
-            error
-        );
     }
 }
