+++ src/artifact_service/handlers.rs
@@ -14,23 +14,18 @@
    limitations under the License.
 */
 
-use super::config::get_config;
-use super::ArtifactManager;
-use super::Hash;
-use super::HashAlgorithm;
-
-use crate::metadata_manager::metadata::Metadata;
-use crate::util::env_util::*;
-use anyhow::{Context, Result};
+use crate::artifact_service::service::{Hash, HashAlgorithm};
+use crate::artifact_service::storage::ArtifactStorage;
+use crate::cli_commands::config::get_config;
+use crate::network::client::{ArtifactType, Client};
+use anyhow::{bail, Context, Result};
 use byte_unit::Byte;
-use lazy_static::lazy_static;
-use log::{debug, error, info};
+use libp2p::PeerId;
+use log::{debug, info};
 use std::collections::HashMap;
 use std::fs::File;
 use std::io::{BufReader, Read};
-use std::panic::UnwindSafe;
 use std::str;
-use std::{fs, panic};
 use sysinfo::{NetworkExt, ProcessExt, System, SystemExt};
 
 //TODO: read from CLI config file
@@ -44,119 +39,100 @@ const DISK_STRESS_WEIGHT: f64 = 0.001_f64;
 //This structure is used as the entries to the quality metrics vector
 //#[derive(Debug, Clone, Copy)]
 
-lazy_static! {
-    pub static ref ARTIFACTS_DIR: String = {
-        let pyrsia_artifact_path = read_var("PYRSIA_ARTIFACT_PATH", "pyrsia");
-        let dev_mode = read_var("DEV_MODE", "off");
-        if dev_mode.to_lowercase() == "on" {
-            log_static_initialization_failure(
-                "Pyrsia Artifact directory",
-                fs::create_dir_all(&pyrsia_artifact_path).with_context(|| {
-                    format!(
-                        "Failed to create artifact manager directory {:?} in dev mode",
-                        pyrsia_artifact_path
-                    )
-                }),
-            );
-        }
-        pyrsia_artifact_path
-    };
-    pub static ref ART_MGR: ArtifactManager = log_static_initialization_failure(
-        "Artifact Manager",
-        ArtifactManager::new(ARTIFACTS_DIR.as_str()),
-    );
-    pub static ref METADATA_MGR: Metadata = log_static_initialization_failure(
-        "Metadata Manager",
-        Metadata::new(ARTIFACTS_DIR.as_str())
-    );
-}
+//get_artifact: given artifact_hash(artifactName) pulls artifact for  artifact_manager and
+//              returns read object to read the bytes of artifact
+pub async fn get_artifact(
+    p2p_client: Client,
+    artifact_storage: &ArtifactStorage,
+    art_hash: &[u8],
+    algorithm: HashAlgorithm,
+) -> Result<Vec<u8>, anyhow::Error> {
+    let hash = Hash::new(algorithm, art_hash)?;
 
-fn log_static_initialization_failure<T: UnwindSafe>(
-    label: &str,
-    result: Result<T, anyhow::Error>,
-) -> T {
-    let panic_wrapper = panic::catch_unwind(|| match result {
-        Ok(unwrapped) => unwrapped,
-        Err(error) => {
-            let msg = format!("Error initializing {}, error is: {}", label, error);
-            error!("{}", msg);
-            panic!("{}", msg)
-        }
-    });
-    match panic_wrapper {
-        Ok(normal) => normal,
-        Err(partially_unwound_panic) => {
-            error!("Initialization of {} panicked!", label);
-            panic::resume_unwind(partially_unwound_panic)
-        }
+    match get_artifact_locally(artifact_storage, &hash) {
+        Ok(blob_content) => Ok(blob_content),
+        Err(_) => get_artifact_from_peers(p2p_client, artifact_storage, &hash).await,
     }
 }
 
-//get_artifact: given artifact_hash(artifactName) pulls artifact for  artifact_manager and
-//              returns read object to read the bytes of artifact
-pub fn get_artifact(art_hash: &[u8], algorithm: HashAlgorithm) -> Result<Vec<u8>, anyhow::Error> {
-    let hash = Hash::new(algorithm, art_hash)?;
-    let result = ART_MGR.pull_artifact(&hash)?;
+fn get_artifact_locally(
+    artifact_storage: &ArtifactStorage,
+    hash: &Hash,
+) -> Result<Vec<u8>, anyhow::Error> {
+    let result = artifact_storage.pull_artifact(hash)?;
     let mut buf_reader: BufReader<File> = BufReader::new(result);
     let mut blob_content = Vec::new();
     buf_reader.read_to_end(&mut blob_content)?;
     Ok(blob_content)
 }
 
-//get_artifact_hashes: retrieve a list of hashes of all artifacts that are stored in
-//                     the artifact_manager
-pub fn get_artifact_hashes() -> Result<Vec<String>, anyhow::Error> {
-    let artifacts = ART_MGR.list_artifacts()?;
-    Ok(artifacts
-        .into_iter()
-        .map(|artifact| {
-            let hash_type = artifact
-                .parent()
-                .unwrap()
-                .file_name()
-                .unwrap()
-                .to_str()
-                .unwrap();
-            let hash_value = artifact.file_name().unwrap().to_str().unwrap();
-            let extension_dot = hash_value.rfind('.').unwrap();
-            format!(
-                "{}:{}",
-                hash_type.to_lowercase(),
-                hash_value.get(0..extension_dot).unwrap()
-            )
-        })
-        .collect())
+async fn get_artifact_from_peers(
+    mut p2p_client: Client,
+    artifact_storage: &ArtifactStorage,
+    hash: &Hash,
+) -> Result<Vec<u8>, anyhow::Error> {
+    let artifact_hash = hex::encode(&hash.bytes);
+
+    let providers = p2p_client
+        .list_providers(ArtifactType::Artifact, (&artifact_hash).into())
+        .await?;
+
+    match p2p_client.get_idle_peer(providers).await? {
+        Some(peer) => get_artifact_from_peer(p2p_client, artifact_storage, &peer, hash).await,
+        None => bail!(
+            "Artifact with hash {} is not available on the p2p network.",
+            artifact_hash
+        ),
+    }
+}
+
+async fn get_artifact_from_peer(
+    mut p2p_client: Client,
+    artifact_storage: &ArtifactStorage,
+    peer_id: &PeerId,
+    hash: &Hash,
+) -> Result<Vec<u8>, anyhow::Error> {
+    let artifact_hash = hex::encode(&hash.bytes);
+
+    let artifact = p2p_client
+        .request_artifact(peer_id, ArtifactType::Artifact, artifact_hash.into())
+        .await?;
+
+    let cursor = Box::new(std::io::Cursor::new(artifact));
+    put_artifact(artifact_storage, hash, cursor)?;
+    get_artifact_locally(artifact_storage, hash)
 }
 
 //put_artifact: given artifact_hash(artifactName) & artifact_path push artifact to artifact_manager
 //              and returns the boolean as true or false if it was able to create or not
 pub fn put_artifact(
-    artifact_hash: &[u8],
+    artifact_storage: &ArtifactStorage,
+    artifact_hash: &Hash,
     art_reader: Box<dyn Read>,
-    algorithm: HashAlgorithm,
-) -> Result<bool, anyhow::Error> {
-    let hash = Hash::new(algorithm, artifact_hash)?;
-    info!("put_artifact hash: {}", hash);
+) -> Result<(), anyhow::Error> {
+    info!("put_artifact hash: {}", artifact_hash);
     let mut buf_reader = BufReader::new(art_reader);
-    ART_MGR
-        .push_artifact(&mut buf_reader, &hash)
+    artifact_storage
+        .push_artifact(&mut buf_reader, artifact_hash)
         .context("Error from put_artifact")
 }
 
-pub fn get_arts_count() -> Result<usize, anyhow::Error> {
-    ART_MGR
+pub fn get_arts_count(artifact_storage: &ArtifactStorage) -> Result<usize, anyhow::Error> {
+    artifact_storage
         .artifacts_count()
         .context("Error while getting artifacts count")
 }
 
-pub fn get_arts_summary() -> Result<HashMap<String, usize>, anyhow::Error> {
-    ART_MGR
+pub fn get_arts_summary(
+    artifact_storage: &ArtifactStorage,
+) -> Result<HashMap<String, usize>, anyhow::Error> {
+    artifact_storage
         .artifacts_count_bydir()
         .context("Error while getting artifacts count")
 }
 
-pub fn get_space_available() -> Result<u64, anyhow::Error> {
-    let disk_used_bytes = ART_MGR.space_used()?;
+pub fn get_space_available(artifact_storage: &ArtifactStorage) -> Result<u64, anyhow::Error> {
+    let disk_used_bytes = artifact_storage.space_used()?;
 
     let mut available_space: u64 = 0;
     let cli_config = get_config().context("Error getting cli config file")?;
@@ -171,8 +147,8 @@ pub fn get_space_available() -> Result<u64, anyhow::Error> {
     Ok(available_space)
 }
 
-pub fn disk_usage() -> Result<f64, anyhow::Error> {
-    let disk_used_bytes = ART_MGR.space_used()?;
+pub fn disk_usage(artifact_storage: &ArtifactStorage) -> Result<f64, anyhow::Error> {
+    let disk_used_bytes = artifact_storage.space_used()?;
     let cli_config = get_config().context("Error getting cli config file")?;
     let total_allocated_size: u64 = Byte::from_str(cli_config.disk_allocated)
         .unwrap()
@@ -238,6 +214,9 @@ mod tests {
     use super::*;
     use anyhow::Context;
     use assay::assay;
+    use futures::channel::mpsc;
+    use futures::executor;
+    use libp2p::identity::Keypair;
     use std::env;
     use std::fs::File;
     use std::path::Path;
@@ -255,7 +234,7 @@ mod tests {
 
     fn tear_down() {
         if Path::new(&env::var("PYRSIA_ARTIFACT_PATH").unwrap()).exists() {
-            fs::remove_dir_all(env::var("PYRSIA_ARTIFACT_PATH").unwrap()).expect(&format!(
+            std::fs::remove_dir_all(env::var("PYRSIA_ARTIFACT_PATH").unwrap()).expect(&format!(
                 "unable to remove test directory {}",
                 env::var("PYRSIA_ARTIFACT_PATH").unwrap()
             ));
@@ -270,17 +249,31 @@ mod tests {
         teardown = tear_down()
         )]
     fn test_put_and_get_artifact() {
-        //put the artifact
-        put_artifact(
-            &VALID_ARTIFACT_HASH,
-            Box::new(get_file_reader()?),
-            HashAlgorithm::SHA256,
-        )
-        .context("Error from put_artifact")?;
+        let artifact_storage = ArtifactStorage::new()?;
 
-        // pull artiafct
-        let file = get_artifact(&VALID_ARTIFACT_HASH, HashAlgorithm::SHA256)
-            .context("Error from get_artifact")?;
+        let (sender, _) = mpsc::channel(1);
+        let p2p_client = Client {
+            sender,
+            local_peer_id: Keypair::generate_ed25519().public().to_peer_id(),
+        };
+
+        let hash = Hash::new(HashAlgorithm::SHA256, &VALID_ARTIFACT_HASH)?;
+        //put the artifact
+        put_artifact(&artifact_storage, &hash, Box::new(get_file_reader()?))
+            .context("Error from put_artifact")?;
+
+        // pull artifact
+        let future = async {
+            get_artifact(
+                p2p_client,
+                &artifact_storage,
+                &VALID_ARTIFACT_HASH,
+                HashAlgorithm::SHA256,
+            )
+            .await
+            .context("Error from get_artifact")
+        };
+        let file = executor::block_on(future)?;
 
         //validate pulled artifact with the actual data
         let mut s = String::new();
@@ -299,34 +292,14 @@ mod tests {
           ("DEV_MODE", "on")
         ]  )]
     fn test_disk_usage() {
-        let usage_pct_before = disk_usage().context("Error from disk_usage")?;
+        let artifact_storage = ArtifactStorage::new()?;
 
-        create_artifact().context("Error creating artifact")?;
+        let usage_pct_before = disk_usage(&artifact_storage).context("Error from disk_usage")?;
 
-        let usage_pct_after = disk_usage().context("Error from disk_usage")?;
+        create_artifact(&artifact_storage).context("Error creating artifact")?;
+
+        let usage_pct_after = disk_usage(&artifact_storage).context("Error from disk_usage")?;
         assert!(usage_pct_before < usage_pct_after);
-    }
-
-    #[assay(
-        env = [
-          ("PYRSIA_ARTIFACT_PATH", "PyrsiaTest"),
-          ("DEV_MODE", "on")
-        ]  )]
-    fn test_get_artifact_hashes_is_empty() {
-        let artifact_hashes = get_artifact_hashes().context("Error from get_artifact_hashes")?;
-        assert!(artifact_hashes.is_empty());
-    }
-
-    #[assay(
-        env = [
-          ("PYRSIA_ARTIFACT_PATH", "PyrsiaTest"),
-          ("DEV_MODE", "on")
-        ]  )]
-    fn test_get_artifact_hashes() {
-        create_artifact().context("Error creating artifact")?;
-
-        let artifact_hashes = get_artifact_hashes().context("Error from get_artifact_hashes")?;
-        assert!(artifact_hashes.len() == 1);
     }
 
     fn get_file_reader() -> Result<File, anyhow::Error> {
@@ -339,14 +312,11 @@ mod tests {
         Ok(reader)
     }
 
-    fn create_artifact() -> Result<(), anyhow::Error> {
+    fn create_artifact(artifact_storage: &ArtifactStorage) -> Result<(), anyhow::Error> {
         let hash = Hash::new(HashAlgorithm::SHA256, &VALID_ARTIFACT_HASH)?;
-        let push_result = ART_MGR
+        artifact_storage
             .push_artifact(&mut get_file_reader()?, &hash)
-            .context("Error while pushing artifact")?;
-
-        assert_eq!(push_result, true);
-        Ok(())
+            .context("Error while pushing artifact")
     }
 
     #[test]
