+++ src/docker.rs
@@ -15,6 +15,5 @@
 */
 
 pub mod constants;
-mod docker_hub_util;
 pub mod error_util;
 pub mod v2;
