+++ src/docker/v2.rs
@@ -16,4 +16,3 @@
 
 pub mod handlers;
 pub mod routes;
-pub mod storage;
