+++ src/node_api.rs
@@ -15,6 +15,5 @@
 */
 
 pub mod handlers;
+pub mod model;
 pub mod routes;
-
-pub use crate::node_manager::handlers::*;
