+++ src/node_api/model.rs
@@ -14,5 +14,4 @@
    limitations under the License.
 */
 
-pub mod blobs;
-pub use crate::artifacts_repository::hash_util::HashAlgorithm;
+pub mod cli;
